Dont know whats get Updated! Nothing is done yet!
simple form are created with a little bit of handling, hopefully sort
it out tomorrow

+++ src/bean/Service.java
@@ -21,6 +21,6 @@ public class Service {
 	@DatabaseField(columnName="S_CHTYPE",canBeNull=false)
 	private String chargeType;
 	@DatabaseField(columnName="S_CHARGE",canBeNull=false)
-	private double charge;
+	private int charge;
 
 }
